email/report-data/requirements.txt
altair==5.5.0
appnope==0.1.4
asn1crypto==1.5.1
astroid==3.3.9
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.37.13
botocore==1.37.13
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
cramjam==2.9.1
debugpy==1.8.13
decorator==5.2.1
dill==0.3.9
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
fastparquet==2024.11.0
Flask==3.1.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
narwhals==1.31.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.1
pluggy==1.5.0
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pydeck==0.9.1
Pygments==2.19.1
pylint==3.3.5
PyMySQL==1.1.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyzmq==26.3.0
redshift-connector==2.1.5
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
s3transfer==0.11.4
scramp==1.4.5
setuptools==76.0.0
six==1.17.0
smmap==5.0.2
soupsieve==2.6
stack-data==0.6.3
streamlit==1.43.2
tenacity==9.0.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
